A generic function to create a data structure for every variable pair in a dataset

Creates a data structure for every variable pair in a dataset.

assoc_tibble(data, measure_type = "?", pair_type = "?")

# S3 method for matrix
assoc_tibble(data, measure_type = "?", pair_type = "?")

# S3 method for data.frame
assoc_tibble(data, measure_type = NA_character_, pair_type = NA_character_)

Arguments

data

A dataframe.

measure_type

a character string indicating the measure of association.

pair_type

a character string specifying the type of variable pair.

Value

A data structure for pairs of variables with a column measure for measure value, measure_type for a type of association measure and pair_type for the variable pair.

Methods (by class)

assoc_tibble(matrix): assoc_tibble method

assoc_tibble(data.frame): assoc_tibble method

Examples

assoc_tibble(cor(iris[,1:4]), measure_type="pearson")
#> # A tibble: 6 × 5
#>   x            y            measure measure_type pair_type
#>   <chr>        <chr>          <dbl> <chr>        <chr>    
#> 1 Sepal.Width  Sepal.Length  -0.118 pearson      ?        
#> 2 Petal.Length Sepal.Length   0.872 pearson      ?        
#> 3 Petal.Width  Sepal.Length   0.818 pearson      ?        
#> 4 Petal.Length Sepal.Width   -0.428 pearson      ?        
#> 5 Petal.Width  Sepal.Width   -0.366 pearson      ?        
#> 6 Petal.Width  Petal.Length   0.963 pearson      ?        
assoc_tibble(iris)
#> # A tibble: 10 × 5
#>    x            y            measure measure_type pair_type
#>    <chr>        <chr>          <dbl> <chr>        <chr>    
#>  1 Sepal.Width  Sepal.Length      NA NA           NA       
#>  2 Petal.Length Sepal.Length      NA NA           NA       
#>  3 Petal.Width  Sepal.Length      NA NA           NA       
#>  4 Species      Sepal.Length      NA NA           NA       
#>  5 Petal.Length Sepal.Width       NA NA           NA       
#>  6 Petal.Width  Sepal.Width       NA NA           NA       
#>  7 Species      Sepal.Width       NA NA           NA       
#>  8 Petal.Width  Petal.Length      NA NA           NA       
#>  9 Species      Petal.Length      NA NA           NA       
#> 10 Species      Petal.Width       NA NA           NA