Calculates multiple association measures

Calculates multiple association measures for every variable pair in a dataset.

calc_assoc_all(
  d,
  measures = c("pearson", "spearman", "kendall", "cancor", "nmi", "dcor", "mic", "ace",
    "polycor", "tau_b", "uncertainty", "gkTau", "gkGamma", "chi"),
  handle.na = T
)

Arguments

d

dataframe

measures

a set of all the measures such as "pearson","spearman","kendall", "cancor","nmi","dcor", "mic", "ace, "polycor", "tau_b", "uncertainty", "gkTau", "gkGamma" and "chi" available in the package. Set to all the measures by default and can be updated to a subset of these measures.

handle.na

If TRUE uses pairwise complete observations to calculate measure of association

Value

tibble of class "multi_pairwise"

Examples

calc_assoc_all(iris)
#> # A tibble: 60 × 5
#>    x            y            measure measure_type pair_type
#>    <chr>        <chr>          <dbl> <chr>        <chr>    
#>  1 Sepal.Width  Sepal.Length  -0.118 pearson      nn       
#>  2 Petal.Length Sepal.Length   0.872 pearson      nn       
#>  3 Petal.Width  Sepal.Length   0.818 pearson      nn       
#>  4 Petal.Length Sepal.Width   -0.428 pearson      nn       
#>  5 Petal.Width  Sepal.Width   -0.366 pearson      nn       
#>  6 Petal.Width  Petal.Length   0.963 pearson      nn       
#>  7 Sepal.Width  Sepal.Length  -0.167 spearman     nn       
#>  8 Petal.Length Sepal.Length   0.882 spearman     nn       
#>  9 Petal.Width  Sepal.Length   0.834 spearman     nn       
#> 10 Petal.Length Sepal.Width   -0.310 spearman     nn       
#> # … with 50 more rows