A generic function to create a symmetric matrix from a tibble of association measures with different classes

Creates a symmetric association matrix from a tibble of association measures with classes such as pairwise, cond_pairwise and multi_pairwise. For an association measure tibble of class multi_pairwise, a matrix is not calculated.

matrix_assoc(assoc, group = NULL)

# S3 method for tbl
matrix_assoc(assoc, group = NULL)

# S3 method for pairwise
matrix_assoc(assoc, group = NULL)

# S3 method for cond_pairwise
matrix_assoc(assoc, group = "overall")

# S3 method for multi_pairwise
matrix_assoc(assoc, group = NULL)

Arguments

assoc

A tibble or dataframe with calculated association measures for variable pairs.

group

A character string specifying the level of the conditioning variable for which a symmetric matrix needs to be create. One of "overall" (default) or a level of conditioning variable.

Value

matrix symmetric matrix of variable pairs with corresponding association measure

Methods (by class)

matrix_assoc(tbl): matrix_assoc method

matrix_assoc(pairwise): matrix_assoc method

matrix_assoc(cond_pairwise): matrix_assoc method

matrix_assoc(multi_pairwise): matrix_assoc method

Examples

# Symmetric matrix for assoc with class `pairwise`
a <- calc_assoc(iris)
matrix_assoc(a)
#>              Sepal.Length Sepal.Width Petal.Length Petal.Width   Species
#> Sepal.Length           NA  -0.1175698    0.8717538   0.8179411 0.7865785
#> Sepal.Width    -0.1175698          NA   -0.4284401  -0.3661259 0.6330741
#> Petal.Length    0.8717538  -0.4284401           NA   0.9628654 0.9702431
#> Petal.Width     0.8179411  -0.3661259    0.9628654          NA 0.9637857
#> Species         0.7865785   0.6330741    0.9702431   0.9637857        NA

# Symmetric matrix for assoc with class `cond_pairwise`
b <- calc_assoc(iris, by="Species")
matrix_assoc(b,group="setosa")
#>              Sepal.Length Sepal.Width Petal.Length Petal.Width
#> Sepal.Length           NA   0.7425467    0.2671758   0.2780984
#> Sepal.Width     0.7425467          NA    0.1777000   0.2327520
#> Petal.Length    0.2671758   0.1777000           NA   0.3316300
#> Petal.Width     0.2780984   0.2327520    0.3316300          NA