Ordering variables

Calculates an ordering for the variables in a dataset.

order_assoc_var(assoc, group_var = group_var)

Arguments

assoc

A tibble with association measures for different variable pairs in the dataset.

group_var

a character vector for grouping variable. One of NULL, "by" or "measure_type"

Value

character vector representing the ordering of the variables

Examples

order_assoc_var(calc_assoc(iris))
#> Error in order_assoc_var(calc_assoc(iris)): promise already under evaluation: recursive default argument reference or earlier problems?
order_assoc_var(calc_assoc(iris),"by")
#> [1] "Species"      "Petal.Length" "Sepal.Width"  "Sepal.Length" "Petal.Width" 
order_assoc_var(calc_assoc_all(iris),"measure_type")
#> [1] "Petal.Width"  "Species"      "Petal.Length" "Sepal.Length" "Sepal.Width"