Uncertainty coefficient

Calculates uncertainty coefficient for every nominal variable pair in a dataset.

tbl_uncertainty(d, handle.na = TRUE, ...)

Arguments

d

A dataframe or tibble

handle.na

If TRUE uses pairwise complete observations to calculate correlation coefficient

...

other arguments

Value

A tibble with every nominal variable pair and uncertainty coefficient value.

Details

The Uncertainty coefficient is calculated using UncertCoef function from the DescTools package.

Examples

tbl_uncertainty(iris)
#> # A tibble: 0 × 5
#> # … with 5 variables: x <chr>, y <chr>, measure <named list>,
#> #   measure_type <chr>, pair_type <chr>