A user friendly function for changing association measures

Creates a tibble for different measures of association for different variable types of a dataset.

update_assoc(
  default = default_assoc(),
  num_pair = NULL,
  num_pair_argList = NULL,
  factor_pair = NULL,
  factor_pair_argList = NULL,
  ordered_pair = NULL,
  ordered_pair_argList = NULL,
  mixed_pair = NULL,
  mixed_pair_argList = NULL,
  ...
)

Arguments

default

default measure functions for different variable pairs. set to default_assoc()

num_pair

a measure(s) function for numeric pair of variables, default is NULL

num_pair_argList

a character string specifying the measure to be calculated using num_pair, default is NULL

factor_pair

a measure(s) function for factor pair of variables, default is NULL

factor_pair_argList

a character string specifying the measure to be calculated using factor_pair, default is NULL

ordered_pair

a measure(s) function for ordered pair of variables, default is NULL

ordered_pair_argList

a character string specifying the measure to be calculated using ordered_pair, default is NULL

mixed_pair

a measure(s) function for mixed pair of variables, default is NULL

mixed_pair_argList

a character string specifying the measure to be calculated using mixed_pair, default is NULL

...

other arguments

Value

tibble

Examples

updated_assoc <- update_assoc(num_pair="tbl_cor", num_pair_argList="spearman",
ordered_pair="tbl_tau",mixed_pair="tbl_nmi",factor_pair="tbl_cancor")
calc_assoc(iris,types=updated_assoc)
#> # A tibble: 10 × 5
#>    x            y            measure measure_type pair_type
#>    <chr>        <chr>          <dbl> <chr>        <chr>    
#>  1 Sepal.Width  Sepal.Length  -0.167 spearman     nn       
#>  2 Petal.Length Sepal.Length   0.882 spearman     nn       
#>  3 Petal.Width  Sepal.Length   0.834 spearman     nn       
#>  4 Species      Sepal.Length   0.487 nmi          nf       
#>  5 Petal.Length Sepal.Width   -0.310 spearman     nn       
#>  6 Petal.Width  Sepal.Width   -0.289 spearman     nn       
#>  7 Species      Sepal.Width    0.261 nmi          nf       
#>  8 Petal.Width  Petal.Length   0.938 spearman     nn       
#>  9 Species      Petal.Length   0.870 nmi          nf       
#> 10 Species      Petal.Width    0.892 nmi          nf