Association/Conditional Association measures for a dataset

Calculates association measures for every variable pair or conditional association measures for every variable pair at different levels of a third (conditional) variable in a dataset.

Usage

calc_assoc(
  d,
  by = NULL,
  types = default_assoc(),
  include.overall = TRUE,
  handle.na = TRUE,
  coerce_types = NULL
)

Arguments

d

dataframe

by

a character string for the name of the conditional variable. Set to NULL by default.

types

a tibble for the measures to be calculated for different variable types. The default is default_assoc() which calculates Pearson's correlation if the variable pair is numeric, Kendall's tau B if variable pair is ordered factor, canonical correlation if one is numeric and other is a factor, and canonical correlation for any other variable pair.

include.overall

Useful during calculation of conditional association measures. If TRUE calculates the overall measure of association for every pair of variable and includes it in the result.

handle.na

If TRUE uses pairwise complete observations to calculate measure of association.

coerce_types

a list specifying the variables that to need to be coerced to different variable types

Value

A tibble with class "pairwise" when by argument is set to NULL. Else a tibble with class "cond_pairwise"

Examples

calc_assoc(iris)
#> # A tibble: 10 × 5
#>    x            y            measure measure_type pair_type
#>    <chr>        <chr>          <dbl> <chr>        <chr>    
#>  1 Sepal.Width  Sepal.Length  -0.118 pearson      nn       
#>  2 Petal.Length Sepal.Length   0.872 pearson      nn       
#>  3 Petal.Width  Sepal.Length   0.818 pearson      nn       
#>  4 Species      Sepal.Length   0.787 cancor       nf       
#>  5 Petal.Length Sepal.Width   -0.428 pearson      nn       
#>  6 Petal.Width  Sepal.Width   -0.366 pearson      nn       
#>  7 Species      Sepal.Width    0.633 cancor       nf       
#>  8 Petal.Width  Petal.Length   0.963 pearson      nn       
#>  9 Species      Petal.Length   0.970 cancor       nf       
#> 10 Species      Petal.Width    0.964 cancor       nf       
calc_assoc(iris, by = "Species")
#> # A tibble: 24 × 6
#>    x            y            measure measure_type by         pair_type
#>    <chr>        <chr>          <dbl> <chr>        <fct>      <chr>    
#>  1 Sepal.Width  Sepal.Length   0.743 pearson      setosa     nn       
#>  2 Petal.Length Sepal.Length   0.267 pearson      setosa     nn       
#>  3 Petal.Width  Sepal.Length   0.278 pearson      setosa     nn       
#>  4 Petal.Length Sepal.Width    0.178 pearson      setosa     nn       
#>  5 Petal.Width  Sepal.Width    0.233 pearson      setosa     nn       
#>  6 Petal.Width  Petal.Length   0.332 pearson      setosa     nn       
#>  7 Sepal.Width  Sepal.Length   0.526 pearson      versicolor nn       
#>  8 Petal.Length Sepal.Length   0.754 pearson      versicolor nn       
#>  9 Petal.Width  Sepal.Length   0.546 pearson      versicolor nn       
#> 10 Petal.Length Sepal.Width    0.561 pearson      versicolor nn       
#> # … with 14 more rows